+++ src/main/java/kitchenpos/menugroup/domain/JpaMenuGroupRepository.java
@@ -1,8 +1,7 @@
-package kitchenpos.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
+package kitchenpos.menugroup.domain;
 
 import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaMenuGroupRepository extends MenuGroupRepository, JpaRepository<MenuGroup, UUID> {
 }
