+++ src/main/java/kitchenpos/menugroup/domain/MenuGroup.java
@@ -1,10 +1,10 @@
-package kitchenpos.domain;
+package kitchenpos.menugroup.domain;
 
+import java.util.UUID;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Id;
 import javax.persistence.Table;
-import java.util.UUID;
 
 @Table(name = "menu_group")
 @Entity
