+++ src/main/java/kitchenpos/menugroup/ui/MenuGroupRestController.java
@@ -1,12 +1,15 @@
-package kitchenpos.ui;
-
-import kitchenpos.application.MenuGroupService;
-import kitchenpos.domain.MenuGroup;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+package kitchenpos.menugroup.ui;
 
 import java.net.URI;
 import java.util.List;
+import kitchenpos.menugroup.application.MenuGroupService;
+import kitchenpos.menugroup.domain.MenuGroup;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RequestMapping("/api/menu-groups")
 @RestController
